format code and add copyright header (#460)

+++ dubbo-multiple-datasource-mybatis-plus/sql/test_order.sql
@@ -13,16 +13,19 @@ File Encoding         : 65001
 Date: 2019-12-08 15:11:11
 */
 
-SET FOREIGN_KEY_CHECKS=0;
+SET
+FOREIGN_KEY_CHECKS=0;
 
 -- ----------------------------
 -- Table structure for orders
 -- ----------------------------
 DROP TABLE IF EXISTS `orders`;
-CREATE TABLE `orders` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `product_id` int(11) DEFAULT NULL,
-  `amount` double(11,2) DEFAULT NULL,
+CREATE TABLE `orders`
+(
+    `id`         int(11) NOT NULL AUTO_INCREMENT,
+    `product_id` int(11) DEFAULT NULL,
+    `amount`     double(11, 2
+) DEFAULT NULL,
   `sum` int(11) DEFAULT NULL,
   `create_time` datetime DEFAULT NULL,
   `replace_time` datetime DEFAULT NULL,
@@ -33,24 +36,26 @@ CREATE TABLE `orders` (
 -- ----------------------------
 -- Records of orders
 -- ----------------------------
-INSERT INTO `orders` VALUES ('2', '1', '5.00', '1', '2019-12-08 15:05:05', '2019-12-08 15:05:05', '1');
+INSERT INTO `orders`
+VALUES ('2', '1', '5.00', '1', '2019-12-08 15:05:05', '2019-12-08 15:05:05', '1');
 
 -- ----------------------------
 -- Table structure for undo_log
 -- ----------------------------
 DROP TABLE IF EXISTS `undo_log`;
-CREATE TABLE `undo_log` (
-  `id` bigint(20) NOT NULL AUTO_INCREMENT,
-  `branch_id` bigint(20) NOT NULL,
-  `xid` varchar(100) NOT NULL,
-  `context` varchar(128) NOT NULL,
-  `rollback_info` longblob NOT NULL,
-  `log_status` int(11) NOT NULL,
-  `log_created` datetime NOT NULL,
-  `log_modified` datetime NOT NULL,
-  `ext` varchar(100) DEFAULT NULL,
-  PRIMARY KEY (`id`),
-  UNIQUE KEY `ux_undo_log` (`xid`,`branch_id`)
+CREATE TABLE `undo_log`
+(
+    `id`            bigint(20) NOT NULL AUTO_INCREMENT,
+    `branch_id`     bigint(20) NOT NULL,
+    `xid`           varchar(100) NOT NULL,
+    `context`       varchar(128) NOT NULL,
+    `rollback_info` longblob     NOT NULL,
+    `log_status`    int(11) NOT NULL,
+    `log_created`   datetime     NOT NULL,
+    `log_modified`  datetime     NOT NULL,
+    `ext`           varchar(100) DEFAULT NULL,
+    PRIMARY KEY (`id`),
+    UNIQUE KEY `ux_undo_log` (`xid`,`branch_id`)
 ) ENGINE=InnoDB AUTO_INCREMENT=3 DEFAULT CHARSET=utf8;
 
 -- ----------------------------
